+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Use authManager for logs and traces. This is creating the secrets per cluster as well as the ingress and http route secrets but those are not used yet.
+
 ## [0.51.2] - 2025-12-08
 
 ### Fixed