+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add support for per-tenant datasources in Grafana based on tenant access types (data vs alerting) to enable fine-grained control over alerting permissions.
 - Use authManager for logs and traces. This is creating the secrets per cluster as well as the ingress and http route secrets but those are not used yet.
 
 ## [0.51.2] - 2025-12-08